Solve day 9 part 1

+++ app/src/main/java/com/gcjensen/aoc2019/Intcode/IntcodeIO.java
@@ -1,7 +1,7 @@
 package com.gcjensen.aoc2019.Intcode;
 
 public interface IntcodeIO {
-    void output(Integer val);
+    void output(long val);
 
-    Integer input();
+    Long input();
 }
