+++ app/src/test/java/com/gcjensen/aoc2019/day5/Day5Test.java
@@ -7,9 +7,9 @@ import static org.junit.Assert.assertEquals;
 import com.gcjensen.aoc2019.Day;
 
 public class Day5Test {
-    Day<Integer> day5 = new Day5();
+    Day<Long> day5 = new Day5();
 
-    private final List<Integer> actualInput = day5.parseInput();
+    private final List<Long> actualInput = day5.parseInput();
 
     @Test public void testPart1() {
         assertEquals(10987514, day5.solvePart1(actualInput));
